Messages, Methods, and Optimization
I've added a new method to TPASection.java to remove all existing incoming TPA requests. It is called when a player disables incoming TPA requests with /tpatoggle.
I've removed an unnecessary variable from ToggledPlayer.java. It did not behave as intended. The check on a players toggled status is done properly through TPASection.